requirements_1.txt
asgiref==3.7.2
Babel==2.12.1
boto3==1.26.160
botocore==1.29.160
# certifi==2022.9.24
charset-normalizer==2.1.1
decorator===5.1.1
Django==4.2.2
django-simple-history==3.1.1
djangorestframework==3.14.0
idna== 3.4
instaloader==4.9.5
Jinja2==  3.1.2
jmespath==1.0.1
lxml== 4.9.2
MarkupSafe==2.1.3
passlib== 1.7.4
paytm-pg==1.1.1
Pillow==  9.5.0
pip==  22.2.2
polib==1.2.0
psutil==  5.9.5
psycopg2-binary==2.9.6
pycryptodome====3.18.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
reportlab==4.0.4
# requests==2.28.1
s3transfer==0.6.1
# setuptools==65.4.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.13
Werkzeug==2.3.6
# wheel==0.37.1